+++ CHANGELOG.md
@@ -9,6 +9,8 @@ XialotEcon Changelog
     - AccountPriorityEvent to determine which account should be used
     - AccountDescriptionEvent to display an account to a player
     - Account is touched whenever instance is downloaded, to track obsolete accounts
+  - Bank
+    - Implemented bank interest (constant ratio or constant difference)
   - Currency
     - Auto-sync among servers
   - Database connection: powered by libasynql
@@ -17,6 +19,9 @@ XialotEcon Changelog
     - Added /list-cache command
   - Player
     - Auto init accounts
+    - /pay command
+    - /balance command
+    - PlayerAccountDefinitionEvent for adding new user-friendly account types using player.defaults config
   - Transaction
   - Util
     - Config parses time durations in given units